Extract download-related methods in a separate class
